Rename PostgresIntegrationTest to IntegrationTest as all integration tests use PostgreSQL
