+++ src/test/java/nl/b3p/tailormap/api/controller/UniqueValuesControllerIntegrationTest.java
@@ -39,7 +39,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
@@ -47,9 +46,8 @@ import org.springframework.test.web.servlet.MvcResult;
 @AutoConfigureMockMvc
 @Execution(ExecutionMode.CONCURRENT)
 @Stopwatch
-@TestPropertySource(properties = {"tailormap-api.unique.use_geotools_unique_function=false"})
-@Order(2)
-class UniqueValuesControllerReConfiguredPostgresIntegrationTest {
+@Order(1)
+class UniqueValuesControllerIntegrationTest {
   private static final String provinciesWFSUrl =
       "/app/default/layer/lyr:pdok-kadaster-bestuurlijkegebieden:Provinciegebied/unique/naam";
   private static final String begroeidterreindeelPostgisUrl =
@@ -96,6 +94,32 @@ class UniqueValuesControllerReConfiguredPostgresIntegrationTest {
 
     assertEquals(values.size(), uniqueValues.size(), "Unique values should be unique");
     assertTrue(uniqueValues.containsAll(Set.of(expected)), "not all values are present");
+  }
+
+  @Test
+  @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
+  void layer_without_featuretype() throws Exception {
+    final String url =
+        apiBasePath
+            + "/app/default/layer/lyr:pdok-kadaster-bestuurlijkegebieden:Gemeentegebied/unique/naam";
+    mockMvc
+        .perform(get(url).accept(MediaType.APPLICATION_JSON).with(setServletPath(url)))
+        .andExpect(status().is4xxClientError())
+        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+        .andExpect(jsonPath("$.message").value("Layer does not have feature type"));
+  }
+
+  @Test
+  @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
+  void attribute_name_required() throws Exception {
+    final String url =
+        apiBasePath
+            + "/app/default/layer/lyr:pdok-kadaster-bestuurlijkegebieden:Gemeentegebied/unique/ ";
+    mockMvc
+        .perform(get(url).accept(MediaType.APPLICATION_JSON).with(setServletPath(url)))
+        .andExpect(status().is4xxClientError())
+        .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+        .andExpect(jsonPath("$.message").value("Attribute name is required"));
   }
 
   @ParameterizedTest(
@@ -253,6 +277,31 @@ class UniqueValuesControllerReConfiguredPostgresIntegrationTest {
         .andExpect(jsonPath("$.values[1]").value(Matchers.containsString("-Holland")));
   }
 
+  /**
+   * Testcase for <a href="https://b3partners.atlassian.net/browse/HTM-492">HTM-492</a> where
+   * Jackson fails to process oracle.sql.TIMESTAMP.
+   *
+   * <p>The exception is: {@code org.springframework.web.util.NestedServletException: Request
+   * processing failed; nested exception is java.lang.ClassCastException: class oracle.sql.TIMESTAMP
+   * cannot be cast to class java.lang.Comparable (oracle.sql.TIMESTAMP is in unnamed module of
+   * loader 'app'; java.lang.Comparable is in module java.base of loader 'bootstrap') }
+   *
+   * <p>For this testcase to go green set the environment variable {@code
+   * -Doracle.jdbc.J2EE13Compliant=true}
+   *
+   * <p>See also:
+   *
+   * <ul>
+   *   <li><a
+   *       href="https://stackoverflow.com/questions/13269564/java-lang-classcastexception-oracle-sql-timestamp-cannot-be-cast-to-java-sql-ti">java.lang.ClassCastException:
+   *       oracle.sql.TIMESTAMP cannot be cast to java.sql.Timestamp</a>
+   *   <li><a
+   *       href="https://docs.oracle.com/en/database/oracle/oracle-database/19/jjdbc/accessing-and-manipulating-Oracle-data.html#GUID-C23007CA-E25D-4747-A3C0-4DE219AF56BD">Accessing
+   *       and Manipulating Oracle Data</a>
+   * </ul>
+   *
+   * @throws Exception if any
+   */
   @Issue("https://b3partners.atlassian.net/browse/HTM-492")
   @Test
   @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
